+++ frontend/packages/data-portal/app/components/Download/APIDownloadTab.tsx
@@ -1,6 +1,6 @@
 import { Callout } from '@czi-sds/components'
 import { useMemo } from 'react'
-import { match } from 'ts-pattern'
+import { match, P } from 'ts-pattern'
 
 import { CopyBox } from 'app/components/CopyBox'
 import { I18n } from 'app/components/I18n'
@@ -14,46 +14,112 @@ export function APIDownloadTab() {
   const { t } = useI18n()
   const { datasetId, tomogramId, tomogramVoxelId, type } =
     useDownloadModalContext()
-  const { downloadConfig } = useDownloadModalQueryParamState()
+  const { annotationId, downloadConfig, fileFormat } =
+    useDownloadModalQueryParamState()
   const { logPlausibleCopyEvent } = useLogPlausibleCopyEvent()
 
-  const { label, resourceId, logType } = useMemo(
+  const downloadFunction = match(fileFormat)
+    .with('mrc', () => 'download_mrcfile')
+    .with('zarr', () => 'download_omezarr')
+    .with('ndjson', () => 'download_ndjson')
+    .otherwise(() => '')
+
+  const { label, content, logType } = useMemo(
     () =>
-      match({ type, downloadConfig })
+      match({ fileFormat, type, downloadConfig })
         .with({ type: 'dataset' }, () => ({
           label: t('datasetId'),
-          resourceId: datasetId,
+          content: datasetId,
           logType: 'dataset-id',
         }))
         .with(
           { type: 'runs', downloadConfig: DownloadConfig.AllAnnotations },
           () => ({
             label: t('voxelSpacingId'),
-            resourceId: tomogramVoxelId,
+            content: tomogramVoxelId,
             logType: 'voxel-spacing-id',
           }),
         )
+        .with(
+          {
+            type: 'runs',
+            downloadConfig: DownloadConfig.Tomogram,
+            fileFormat: P.string,
+          },
+          () => ({
+            label: t('copyApiCodeSnippet'),
+            content: (
+              <>
+                from cryoet_data_portal import Client, Tomogram
+                <br />
+                client = Client()
+                <br />
+                tomogram = Tomogram.get_by_id(client, {tomogramId})
+                <br />
+                tomogram.{downloadFunction}()
+              </>
+            ),
+            logType: 'tomogram-code-snippet',
+          }),
+        )
+        .with({ type: 'annotation', fileFormat: P.string }, () => ({
+          label: t('copyApiCodeSnippet'),
+          content: (
+            <>
+              from cryoet_data_portal import Client, Annotation
+              <br />
+              client = Client()
+              <br />
+              annotation = Annotation.get_by_id(client, {annotationId})
+              <br />
+              annotation.download(format=&quot;{fileFormat}&quot;)
+            </>
+          ),
+          logType: 'annotation-code-snippet',
+        }))
         .otherwise(() => ({
           // no idea why this is throwing an error
           // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
           label: t('tomogramId'),
-          resourceId: tomogramId,
+          content: tomogramId,
           logType: 'tomogram-id',
         })),
-    [datasetId, downloadConfig, t, tomogramId, tomogramVoxelId, type],
+    [
+      annotationId,
+      datasetId,
+      downloadConfig,
+      downloadFunction,
+      fileFormat,
+      t,
+      tomogramId,
+      tomogramVoxelId,
+      type,
+    ],
   )
 
   return (
     <div className="pt-sds-xl">
-      <Callout className="!w-full" intent="info">
-        <I18n i18nKey="preferToDownloadViaApi" />
+      <Callout className="!w-full !mt-0" intent="info">
+        <I18n
+          i18nKey={
+            fileFormat ? 'preferToDownloadViaApiCode' : 'preferToDownloadViaApi'
+          }
+          values={{
+            url: `https://chanzuckerberg.github.io/cryoet-data-portal/python-api.html#${
+              annotationId ? 'annotation' : 'tomogram'
+            }`,
+          }}
+          tOptions={{
+            interpolation: { escapeValue: false },
+          }}
+        />
       </Callout>
 
       <CopyBox
         className="mt-sds-l"
-        content={resourceId}
+        content={content}
         title={label}
-        onCopy={() => logPlausibleCopyEvent(logType, String(resourceId))}
+        onCopy={() => logPlausibleCopyEvent(logType, String(content))}
       />
     </div>
   )