+++ frontend/packages/data-portal/app/components/Download/DownloadModal.tsx
@@ -1,4 +1,4 @@
-import { Button, Dialog, DialogActions, Icon } from '@czi-sds/components'
+import { Button, ButtonIcon, Dialog, DialogActions } from '@czi-sds/components'
 import { useCallback, useMemo } from 'react'
 import { match } from 'ts-pattern'
 
@@ -9,6 +9,7 @@ import {
 } from 'app/context/DownloadModal.context'
 import { useDownloadModalQueryParamState } from 'app/hooks/useDownloadModalQueryParamState'
 import { useI18n } from 'app/hooks/useI18n'
+import { PlausibleDownloadModalPayload } from 'app/hooks/usePlausible'
 import { DownloadStep } from 'app/types/download'
 
 import { ConfigureDownloadContent } from './ConfigureDownloadContent'
@@ -17,16 +18,18 @@ import { DownloadOptionsContent } from './DownloadOptionsContent'
 function DownloadModalContent() {
   const { t } = useI18n()
   const {
+    annotationId,
     closeDownloadModal,
     configureDownload,
     downloadConfig,
     downloadStep,
+    fileFormat,
     goBackToConfigure,
     isModalOpen,
   } = useDownloadModalQueryParamState()
   const { datasetId, runId, type, fileSize } = useDownloadModalContext()
 
-  const plausiblePayload = useMemo(
+  const plausiblePayload = useMemo<PlausibleDownloadModalPayload>(
     () => ({
       datasetId,
       fileSize,
@@ -40,39 +43,46 @@ function DownloadModalContent() {
     [closeDownloadModal, plausiblePayload],
   )
 
-  const modalData = useMemo(
-    () =>
-      match({ downloadStep, type })
-        .with(
-          { type: 'dataset' },
-          { type: 'runs', downloadStep: DownloadStep.Download },
-          () => ({
-            buttonDisabled: false,
-            buttonText: t('close'),
-            content: <DownloadOptionsContent />,
-            onClick: closeModal,
-            showBackButton: type === 'runs',
-            title: t('downloadOptions'),
-          }),
-        )
-        .otherwise(() => ({
-          buttonDisabled: !downloadConfig,
-          buttonText: t('next'),
-          content: <ConfigureDownloadContent />,
-          onClick: () => configureDownload(plausiblePayload),
-          showBackButton: false,
-          title: t('configureDownload'),
-        })),
-    [
-      closeModal,
-      configureDownload,
-      downloadConfig,
-      downloadStep,
-      plausiblePayload,
-      t,
-      type,
-    ],
-  )
+  const modalData = useMemo(() => {
+    const hasMultipleSteps = ['runs', 'annotation'].includes(type)
+
+    return match({ downloadStep, type })
+      .with(
+        { type: 'dataset' },
+        { type: 'runs', downloadStep: DownloadStep.Download },
+        { type: 'annotation', downloadStep: DownloadStep.Download },
+        () => ({
+          buttonDisabled: false,
+          buttonText: t('close'),
+          content: <DownloadOptionsContent />,
+          onClick: closeModal,
+          showBackButton: hasMultipleSteps,
+          subtitle: hasMultipleSteps
+            ? t('stepCount', { count: 2, max: 2 })
+            : null,
+          title: t('downloadOptions'),
+        }),
+      )
+      .otherwise(() => ({
+        buttonDisabled: annotationId ? !fileFormat : !downloadConfig,
+        buttonText: t('next'),
+        content: <ConfigureDownloadContent />,
+        onClick: () => configureDownload(plausiblePayload),
+        showBackButton: false,
+        subtitle: t('stepCount', { count: 1, max: 2 }),
+        title: t('configureDownload'),
+      }))
+  }, [
+    annotationId,
+    closeModal,
+    configureDownload,
+    downloadConfig,
+    downloadStep,
+    fileFormat,
+    plausiblePayload,
+    t,
+    type,
+  ])
 
   return (
     <Dialog
@@ -84,18 +94,25 @@ function DownloadModalContent() {
       canClickOutsideClose={false}
     >
       <div className="flex justify-between">
-        <h2 className="text-sds-header-xl leading-sds-header-xl font-semibold pt-4">
-          {modalData.title}
-        </h2>
+        <div className="flex-col gap-sds-xxs py-sds-xxs">
+          {modalData.subtitle && (
+            <p className="text-sds-gray-500 text-sds-body-xs leading-sds-body-xs">
+              {modalData.subtitle}
+            </p>
+          )}
+          <h2 className="text-sds-header-xl leading-sds-header-xl font-semibold">
+            {modalData.title}
+          </h2>
+        </div>
 
-        <button onClick={closeModal} type="button">
-          <Icon
+        <div>
+          <ButtonIcon
+            onClick={closeModal}
+            sdsSize="medium"
+            sdsType="tertiary"
             sdsIcon="xMark"
-            sdsSize="s"
-            sdsType="iconButton"
-            className="!fill-sds-gray-500"
           />
-        </button>
+        </div>
       </div>
 
       {modalData.content}
@@ -105,7 +122,7 @@ function DownloadModalContent() {
           <Button
             onClick={goBackToConfigure}
             sdsStyle="rounded"
-            sdsType="primary"
+            sdsType="secondary"
           >
             {t('back')}
           </Button>