+++ frontend/packages/data-portal/app/components/Run/AnnotationTable.tsx
@@ -9,6 +9,7 @@ import { I18n } from 'app/components/I18n'
 import { CellHeader, PageTable, TableCell } from 'app/components/Table'
 import { Tooltip } from 'app/components/Tooltip'
 import { MAX_PER_PAGE } from 'app/constants/pagination'
+import { AnnotationTableWidths } from 'app/constants/table'
 import { useDownloadModalQueryParamState } from 'app/hooks/useDownloadModalQueryParamState'
 import { useI18n } from 'app/hooks/useI18n'
 import { useIsLoading } from 'app/hooks/useIsLoading'
@@ -53,6 +54,24 @@ function ConfidenceValue({ value }: { value: number }) {
 
 const MAX_AUTHORS = 2
 
+type MethodTypeLabels = {
+  automated: I18nKeys
+  hybrid: I18nKeys
+  manual: I18nKeys
+}
+
+const methodLabels: MethodTypeLabels = {
+  automated: 'automated',
+  hybrid: 'hybrid',
+  manual: 'manual',
+}
+
+const methodTooltipLabels: MethodTypeLabels = {
+  automated: 'methodTypeAutomated',
+  hybrid: 'methodTypeHybrid',
+  manual: 'methodTypeManual',
+}
+
 export function AnnotationTable() {
   const { isLoadingDebounced } = useIsLoading()
   const { run } = useRunById()
@@ -60,7 +79,7 @@ export function AnnotationTable() {
   const { setActiveAnnotation } = useAnnotation()
   const { t } = useI18n()
 
-  const { openTomogramDownloadModal } = useDownloadModalQueryParamState()
+  const { openAnnotationDownloadModal } = useDownloadModalQueryParamState()
 
   const openAnnotationDrawer = useCallback(
     (annotation: Annotation) => {
@@ -71,6 +90,7 @@ export function AnnotationTable() {
   )
 
   const showMethodType = useFeatureFlag('methodType')
+  const showAnnotationDownload = useFeatureFlag('downloadSingleAnnotation')
 
   const columns = useMemo(() => {
     const columnHelper = createColumnHelper<Annotation>()
@@ -90,8 +110,8 @@ export function AnnotationTable() {
         header: () => (
           <CellHeader
             horizontalAlign="right"
-            hideSortIcon
             tooltip={tooltipI18nKey ? <I18n i18nKey={tooltipI18nKey} /> : null}
+            width={AnnotationTableWidths.confidenceCell}
             {...cellHeaderProps}
           >
             {header}
@@ -101,7 +121,10 @@ export function AnnotationTable() {
           const value = getValue() as number | null
 
           return (
-            <TableCell horizontalAlign="right" minWidth={81} maxWidth={120}>
+            <TableCell
+              horizontalAlign="right"
+              width={AnnotationTableWidths.confidenceCell}
+            >
               {typeof value === 'number' ? (
                 <ConfidenceValue value={value} />
               ) : (
@@ -117,12 +140,17 @@ export function AnnotationTable() {
 
     return [
       columnHelper.accessor('id', {
-        header: t('annotationId'),
+        header: () => (
+          <CellHeader width={AnnotationTableWidths.id}>
+            {t('annotationId')}
+          </CellHeader>
+        ),
+
         cell: ({ row: { original: annotation } }) => (
           <TableCell
             className="flex flex-col gap-sds-xxxs !items-start"
-            minWidth={250}
             renderLoadingSkeleton={false}
+            width={AnnotationTableWidths.id}
           >
             <div className="flex gap-sds-xs items-center">
               <p
@@ -187,13 +215,16 @@ export function AnnotationTable() {
 
       columnHelper.accessor('deposition_date', {
         header: () => (
-          <CellHeader hideSortIcon className="whitespace-nowrap text-ellipsis">
+          <CellHeader
+            className="whitespace-nowrap text-ellipsis"
+            width={AnnotationTableWidths.depositionDate}
+          >
             {t('depositionDate')}
           </CellHeader>
         ),
 
         cell: ({ getValue }) => (
-          <TableCell minWidth={91} maxWidth={120}>
+          <TableCell width={AnnotationTableWidths.depositionDate}>
             <div className="line-clamp-2 text-ellipsis capitalize">
               {getValue()}
             </div>
@@ -202,9 +233,14 @@ export function AnnotationTable() {
       }),
 
       columnHelper.accessor('object_name', {
-        header: t('objectName'),
+        header: () => (
+          <CellHeader width={AnnotationTableWidths.objectName}>
+            {t('objectName')}
+          </CellHeader>
+        ),
+
         cell: ({ getValue }) => (
-          <TableCell minWidth={120} maxWidth={250}>
+          <TableCell width={AnnotationTableWidths.objectName}>
             <div className="line-clamp-2 text-ellipsis capitalize">
               {getValue()}
             </div>
@@ -214,10 +250,15 @@ export function AnnotationTable() {
 
       columnHelper.accessor('files', {
         id: 'shape-type',
-        header: t('objectShapeType'),
+
+        header: () => (
+          <CellHeader width={AnnotationTableWidths.files}>
+            {t('objectShapeType')}
+          </CellHeader>
+        ),
 
         cell: ({ getValue }) => (
-          <TableCell minWidth={100} maxWidth={150}>
+          <TableCell width={AnnotationTableWidths.files}>
             {getValue().at(0)?.shape_type ?? '--'}
           </TableCell>
         ),
@@ -229,23 +270,39 @@ export function AnnotationTable() {
               id: 'method-type',
 
               header: () => (
-                <CellHeader className="whitespace-nowrap" hideSortIcon>
+                <CellHeader
+                  className="whitespace-nowrap"
+                  tooltip={<I18n i18nKey="methodTypeInfo" />}
+                  width={AnnotationTableWidths.methodType}
+                >
                   {t('methodType')}
                 </CellHeader>
               ),
 
-              cell: ({ row: { original: annotation } }) => (
-                <TableCell minWidth={81} maxWidth={120}>
-                  {/* convert to link when activate annotation state is moved to URL */}
-                  <button
-                    className="text-sds-primary-400 text-sds-header-s leading-sds-header-s"
-                    onClick={() => openAnnotationDrawer(annotation)}
-                    type="button"
+              cell: ({ row: { original: annotation } }) => {
+                // TODO: hook up backend to this when implemented
+                const methodType = 'automated'
+
+                return (
+                  <TableCell
+                    width={AnnotationTableWidths.methodType}
+                    tooltip={<I18n i18nKey={methodTooltipLabels[methodType]} />}
+                    tooltipProps={{ placement: 'top' }}
                   >
-                    {t('automated')}
-                  </button>
-                </TableCell>
-              ),
+                    {/* convert to link when activate annotation state is moved to URL */}
+                    <button
+                      className={cnsNoMerge(
+                        'text-sds-primary-400 text-sds-header-s leading-sds-header-s',
+                        'hover:underline decoration-dashed decoration-1 underline-offset-4',
+                      )}
+                      onClick={() => openAnnotationDrawer(annotation)}
+                      type="button"
+                    >
+                      {t(methodLabels[methodType])}
+                    </button>
+                  </TableCell>
+                )
+              },
             }),
           ]
         : []),
@@ -273,9 +330,10 @@ export function AnnotationTable() {
       columnHelper.display({
         id: 'annotation-actions',
         // Render empty cell header so that it doesn't break the table layout
-        header: () => <CellHeader hideSortIcon>{null}</CellHeader>,
+        header: () => <CellHeader width={AnnotationTableWidths.actions} />,
+
         cell: ({ row: { original: annotation } }) => (
-          <TableCell minWidth={120} maxWidth={120}>
+          <TableCell width={AnnotationTableWidths.actions}>
             <div className="flex flex-col gap-sds-xs">
               <Button
                 sdsType="primary"
@@ -288,33 +346,39 @@ export function AnnotationTable() {
                 <span>{t('moreInfo')}</span>
               </Button>
 
-              <Button
-                sdsType="primary"
-                sdsStyle="minimal"
-                onClick={() =>
-                  openTomogramDownloadModal({
-                    datasetId: run.dataset.id,
-                    runId: run.id,
-                  })
-                }
-                startIcon={
-                  <Icon sdsIcon="download" sdsSize="s" sdsType="button" />
-                }
-              >
-                {t('download')}
-              </Button>
+              {showAnnotationDownload && (
+                <Button
+                  sdsType="primary"
+                  sdsStyle="minimal"
+                  onClick={() =>
+                    openAnnotationDownloadModal({
+                      datasetId: run.dataset.id,
+                      runId: run.id,
+                      annotationId: annotation.id,
+                      // FIXME: are we supposed to only access the 1st option? (this is how it is done elsewhere)
+                      objectShapeType: annotation.files[0].shape_type,
+                    })
+                  }
+                  startIcon={
+                    <Icon sdsIcon="download" sdsSize="s" sdsType="button" />
+                  }
+                >
+                  {t('download')}
+                </Button>
+              )}
             </div>
           </TableCell>
         ),
       }),
     ] as ColumnDef<Annotation>[]
   }, [
+    t,
+    showMethodType,
     openAnnotationDrawer,
-    openTomogramDownloadModal,
+    showAnnotationDownload,
+    openAnnotationDownloadModal,
     run.dataset.id,
     run.id,
-    showMethodType,
-    t,
   ])
 
   const annotations = useMemo<Annotation[]>(