+++ frontend/packages/data-portal/app/context/DownloadModal.context.ts
@@ -1,24 +1,26 @@
 import { createContext, useContext } from 'react'
 
 import { GetRunByIdQuery } from 'app/__generated__/graphql'
+import { Annotation } from 'app/state/annotation'
 
-export type DownloadModalType = 'dataset' | 'runs'
+export type DownloadModalType = 'dataset' | 'runs' | 'annotation'
 
 export type TomogramResolution =
   GetRunByIdQuery['runs'][number]['tomogram_stats'][number]['tomogram_resolutions'][number]
 
 export interface DownloadModalContextValue {
+  activeAnnotation?: Annotation | null
+  activeTomogramResolution?: TomogramResolution | null
   allTomogramProcessing?: string[]
   allTomogramResolutions?: TomogramResolution[]
   datasetId?: number
   datasetTitle?: string
   fileSize?: number
   httpsPath?: string
+  objectName?: string
   runId?: number
   runName?: string
-  s3DatasetPrefix?: string
-  s3TomogramPrefix?: string
-  s3TomogramVoxelPrefix?: string
+  s3Path?: string
   tomogramId?: number
   tomogramVoxelId?: number
   type: DownloadModalType