+++ frontend/packages/data-portal/app/graphql/getRunById.server.ts
@@ -143,6 +143,7 @@ const GET_RUN_BY_ID_QUERY = gql(`
           release_date
 
           files {
+            format
             https_path
             s3_path
             shape_type
@@ -151,6 +152,8 @@ const GET_RUN_BY_ID_QUERY = gql(`
           authors(order_by: { primary_annotator_status: desc }) {
             name
             primary_annotator_status
+            corresponding_author_status
+            email
           }
 
           author_affiliations: authors(distinct_on: affiliation_name) {
@@ -185,6 +188,7 @@ const GET_RUN_BY_ID_QUERY = gql(`
           id
           processing
           s3_mrc_scale0
+          s3_omezarr_dir
           size_x
           size_y
           size_z