+++ frontend/packages/data-portal/app/hooks/useDownloadModalQueryParamState.ts
@@ -1,4 +1,5 @@
 import { useCallback, useMemo } from 'react'
+import { match, P } from 'ts-pattern'
 
 import { QueryParams } from 'app/constants/query'
 import { DownloadConfig, DownloadStep, DownloadTab } from 'app/types/download'
@@ -33,6 +34,14 @@ export function useDownloadModalQueryParamState() {
     QueryParams.TomogramSampling,
   )
 
+  const [annotationId, setAnnotationId] = useQueryParam<string>(
+    QueryParams.AnnotationId,
+  )
+
+  const [objectShapeType, setObjectShapeType] = useQueryParam<string>(
+    QueryParams.ObjectShapeType,
+  )
+
   const [fileFormat, setFileFormat] = useQueryParam<string>(
     QueryParams.FileFormat,
   )
@@ -43,6 +52,8 @@ export function useDownloadModalQueryParamState() {
     [QueryParams.DownloadTab]: stringParam<DownloadTab>(),
     [QueryParams.TomogramProcessing]: stringParam(),
     [QueryParams.TomogramSampling]: stringParam(),
+    [QueryParams.AnnotationId]: stringParam(),
+    [QueryParams.ObjectShapeType]: stringParam(),
     [QueryParams.FileFormat]: stringParam(),
   })
 
@@ -59,14 +70,20 @@ export function useDownloadModalQueryParamState() {
         runId,
         tomogramProcessing: rest.tomogramProcessing ?? tomogramProcessing,
         tomogramSampling: rest.tomogramSampling ?? tomogramSampling,
+        annotationId: rest.annotationId ?? annotationId,
+        objectShapeType: rest.objectShapeType ?? objectShapeType,
         step: rest.step ?? downloadStep ?? DownloadStep.Download,
         config: rest.config ?? downloadConfig,
         tab: rest.tab ?? downloadTab,
+        fileFormat: rest.fileFormat ?? fileFormat,
       }) as PlausibleDownloadModalPayload,
     [
+      annotationId,
       downloadConfig,
       downloadStep,
       downloadTab,
+      fileFormat,
+      objectShapeType,
       tomogramProcessing,
       tomogramSampling,
     ],
@@ -110,6 +127,25 @@ export function useDownloadModalQueryParamState() {
     [getPlausiblePayload, plausible, setDownloadStep],
   )
 
+  const openAnnotationDownloadModal = useCallback(
+    (payload: PlausibleDownloadModalPayload) => {
+      plausible(
+        Events.OpenDownloadModal,
+        getPlausiblePayload({
+          ...payload,
+          step: DownloadStep.Configure,
+        }),
+      )
+
+      setDownloadParams({
+        [QueryParams.DownloadStep]: DownloadStep.Configure,
+        [QueryParams.AnnotationId]: String(payload.annotationId),
+        [QueryParams.ObjectShapeType]: payload.objectShapeType,
+      })
+    },
+    [getPlausiblePayload, plausible, setDownloadParams],
+  )
+
   const closeDownloadModal = useCallback(
     (payload: PlausibleDownloadModalPayload) => {
       plausible(Events.CloseDownloadModal, getPlausiblePayload(payload))
@@ -123,10 +159,15 @@ export function useDownloadModalQueryParamState() {
       plausible(Events.ClickNextToDownloadOptions, getPlausiblePayload(payload))
 
       setDownloadParams((prev) => ({
-        [QueryParams.DownloadTab]:
-          prev?.[QueryParams.DownloadConfig] === DownloadConfig.Tomogram
-            ? DownloadTab.Download
-            : DownloadTab.AWS,
+        [QueryParams.DownloadTab]: match({
+          annotationId: prev?.[QueryParams.AnnotationId],
+          fileFormat: prev?.[QueryParams.FileFormat],
+        })
+          .with(
+            { fileFormat: P.string.and(P.not('zarr')) },
+            () => DownloadTab.Download,
+          )
+          .otherwise(() => DownloadTab.AWS),
 
         [QueryParams.DownloadStep]: DownloadStep.Download,
       }))
@@ -175,6 +216,7 @@ export function useDownloadModalQueryParamState() {
   )
 
   return {
+    annotationId,
     closeDownloadModal,
     configureDownload,
     downloadConfig,
@@ -184,14 +226,18 @@ export function useDownloadModalQueryParamState() {
     getPlausiblePayload,
     goBackToConfigure,
     isModalOpen,
+    objectShapeType,
+    openAnnotationDownloadModal,
     openDatasetDownloadModal,
     openTomogramDownloadModal,
     setAllAnnotationsConfig,
+    setAnnotationId,
     setDownloadConfig,
     setDownloadParams,
     setDownloadStep,
     setDownloadTab,
     setFileFormat,
+    setObjectShapeType,
     setTomogramConfig,
     setTomogramProcessing,
     setTomogramSampling,